+++ changelog.md
@@ -1,7 +1,8 @@
 # unreleased
 
+* Move `Data.Text.Utf16.Rope.Mixed` module to `Data.Text.Mixed.Rope`. `Data.Text.Utf16.Rope.Mixed` re-exports `Data.Text.Mixed.Rope` for legacy clients.
 * Add `Data.Text.Utf8.Lines` and `Data.Text.Utf8.Rope` modules for ropes indexed by UTF-8 code units.
-* Add UTF-8 indexing functionality to `Data.Text.Utf16.Rope.Mixed`.
+* Add UTF-8 indexing functionality to `Data.Text.Mixed.Rope`.
 * The metrics stored internally in the rope nodes has changed, which should improve performance by making some re-measuring redundant. As a consequence, the time complexity of `Data.Text.Rope.lengthAsPosition` is now linear in the length of the last line.
 
 # 0.2
